backend/requirements.txt
# PriceMon Backend Requirements
# Python 3.10+

# Core
setuptools>=69.0.0

# Core Django
Django==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
# SQLite comes with Python (default for development)
# PostgreSQL adapter for production
psycopg2-binary==2.9.9

# CORS handling
django-cors-headers==4.3.1

# Rate limiting
django-ratelimit==4.1.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Production server (WSGI)
gunicorn==21.2.0

# Development tools
django-extensions==3.2.3

# Optional: If you want better admin interface
# django-admin-interface==0.26.0
